Add `bash` to image
In an effort to work around the fact that `CIRCLE_COMPARE_URL` is not
consistently being set we are using a third party orb to set the
variable. The script that this orb is running uses syntax that does not
appear to be compatible with `ash`.

`ash` related issue: https://github.com/iynere/compare-url/issues/18